pyproject.toml
[project]
name = "mcp-daktela"
version = "1.0.0"
description = "Read-only MCP server for the Daktela contact center REST API"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0,<3",
    "httpx>=0.27",
    "PyJWT>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
    "ruff>=0.8",
]

[project.scripts]
mcp-daktela = "mcp_daktela.server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 99
